注释 循环引用

+++ src/finance/models/order/detail.go
@@ -5,8 +5,8 @@ type FinanceOrderDetail struct {
 	Order    FinanceOrder `gorm:"AssociationForeignKey:OrderId" json:"-"`
 	OrderId  uint         `json:"-" gorm:"COMMENT:'订单编号'"`
 	Name     string       `json:"name" gorm:"COMMENT:'产品名'"`
-	Quantity int          `json:"quantity" gorm:"COMMENT:'数量'"`
+	Quantity float64      `json:"quantity" gorm:"COMMENT:'数量'"`
 	Unit     int          `json:"unit" gorm:"COMMENT:'单位'"`
-	Price    int          `json:"price" gorm:"COMMENT:'价格'"`
-	Measure  int          `json:"measure" gorm:"COMMENT:'计量单位值,测量值'"`
+	Price    float64      `json:"price" gorm:"COMMENT:'价格'"`
+	Measure  float64      `json:"measure" gorm:"COMMENT:'计量单位值,测量值'"`
 }
